refactor solver.rs as a module
	new file:   src/solver/analyze.rs
	new file:   src/solver/build.rs
	new file:   src/solver/mod.rs
	renamed:    src/solver.rs -> src/solver/search.rs
